Server/pipeline/pixiv-scraper/src/requirements.txt
selenium
pyvirtualdisplay
requests
beautifulsoup4
opencv-python
google-cloud-vision
google-cloud-storage